config: Add `--sqs-endpoint` option
Provide an option to set the _advertised_ endpoint for the SQS queue
URLs. The official AWS client for Java will connect to the queue url as
advertised by the server, regardless of what the `endpoint` is
configured to in the top level AWS client.

When running SmoothMQ behind a load balancer, this should be the public
endpoint of the server.